Refactored to clean up code
The desktop and mobile renders were handled in different places; this code cleanup separates the concerns and consolidates the calculations.
